nested_sampling: callback
nested_sampling( ... Emin ...) does a series of medium-size nested
sampling "batches", and checks after each batch if it's reached the
desired minimum energy. The calling function may want to do something
after eatch batch---e.g. write the resulting data to a file, so it can
be pulled. The "callback" argument lets it do that.